# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Oral Care AI API
  description: |
    REST API for the Oral Care AI healthcare platform.
    Supports user authentication, dental scan uploads, AI analysis, and dentist reviews.
  version: 1.0.0
  contact:
    email: api@oralcare.ai
servers:
  - url: http://localhost/oral-care-ai/backend-php/api
    description: Local XAMPP development
  - url: https://us-central1-oralcare-ai.cloudfunctions.net/api
    description: Firebase Cloud Functions
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, dentist, admin]
        phoneLastFour:
          type: string
          maxLength: 4
        profileImageUrl:
          type: string
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Scan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        scanType:
          type: string
          enum: [basic_rgb, advanced_spectral]
        thumbnailUrl:
          type: string
        status:
          type: string
          enum: [uploaded, processing, analyzed, failed, archived]
        captureDevice:
          type: string
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
    AnalysisResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scanId:
          type: string
          format: uuid
        modelType:
          type: string
          enum: [basic_rgb, advanced_spectral]
        modelVersion:
          type: string
        overallScore:
          type: number
          minimum: 0
          maximum: 100
        confidenceScore:
          type: number
          minimum: 0
          maximum: 1
        findings:
          type: array
          items:
            $ref: '#/components/schemas/Finding'
        riskAreas:
          type: array
          items:
            $ref: '#/components/schemas/RiskArea'
        recommendations:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    Finding:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [none, minimal, mild, moderate, severe]
        location:
          type: string
        confidence:
          type: number
    RiskArea:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
        width:
          type: integer
        height:
          type: integer
        label:
          type: string
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                profileImageUrl:
                  type: string
      responses:
        '200':
          description: Profile updated

  # ==================== SCANS ====================
  /scans:
    get:
      tags: [Scans]
      summary: List user's scans
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [uploaded, processing, analyzed, failed, archived]
        - name: scanType
          in: query
          schema:
            type: string
            enum: [basic_rgb, advanced_spectral]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scan'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Scans]
      summary: Upload new scan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, scanType]
              properties:
                image:
                  type: string
                  format: binary
                scanType:
                  type: string
                  enum: [basic_rgb, advanced_spectral]
                captureDevice:
                  type: string
      responses:
        '201':
          description: Scan uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          description: Invalid file or parameters
...
  /scans/{scanId}/analysis:
    get:
      tags: [Analysis]
      summary: Get analysis results for scan
      security:
        - BearerAuth: []
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '404':
          description: No analysis found

  # ==================== DENTISTS ====================
...